test/logic: fix `LogicTest#assertCommandBehavior(...)` method
We trigger saving of data to storage file only if the user command
mutates Address Book data.

However, this caused unit tests in `LogicTest` class to fail.
The code causing this behaviour is in
`LogicTest#assertCommandBehavior(...)`:

    assertEquals(expectedAddressBook, addressBook);
    assertEquals(lastShownList, logic.getLastShownList());
    assertEquals(addressBook, saveFile.load());

As the storage file is only saved when the user command mutates,
commands that do not trigger `addressBook` to save will cause this
`assertEquals()` check to fail:

    assertEquals(addressBook, saveFile.load());

Let's fix the faulty unit tests by saving `addressBook` before the third
`assertEquals()` check.