UserGuide: update "Saving the data" section
We trigger saving of data to storage file only if the user command
mutates Address Book data.

However, this behavior is briefly documented in the User Guide only.
Users are not informed exactly which commands trigger saving of data to
storage file, and when commands don't do so.

Let's update the "Saving the data" section of the User Guide to include
a detailed listing of behavior that do not cause the Address book data
to be changed and hence do not trigger saving of data to the storage
file.